Rename configurable to preferences
