+++ packages/ra-core/src/index.ts
@@ -1,7 +1,7 @@
 export * from './auth';
 export * from './controller';
 export * from './core';
-export * from './configurable';
+export * from './preferences';
 export * from './dataProvider';
 export * from './export';
 export * from './form';
