+++ packages/ra-core/src/preferences/useRenderTemplate.ts
