+++ packages/ra-core/src/preferences/useSetInspectorTitle.ts
