+++ packages/ra-ui-materialui/src/preferences/Inspector.tsx
