+++ packages/ra-ui-materialui/src/preferences/InspectorButton.tsx
