+++ packages/ra-ui-materialui/src/preferences/InspectorRoot.tsx
